Recover from device failures
Sometimes the device fails to work on the first few reads after it is
connected and powered on. Resetting the device on such errors eventually
allows it to continue.

This seems quite brutish. We might need to rate limit or put a hard limit on
how many times this should occur.