Parse raw readings into Measurement type
So that we can start to make sense of them. Add tests based on fixture data
from `dymodump`. It seems that it makes sense for most of these fields to be
`int8`, except for the weights which should never be negative and would
otherwise wrap.

Most of this is based on the reverse engineering here:

- http://steventsnyder.com/reading-a-dymo-usb-scale-using-python/