Automatically convert from ounces to grams
Rather than raising an error when the scale is in the wrong mode. Per the
comment on the constant, we lose a small amount of precision when
converting, but this shouldn't matter. I might add a boolean flag to
optionally disable conversion at a later date.

Although the original comment for `ScaleFacter` said that it only applies to
ounces mode, it doesn't matter if we apply it in grams mode because it
should be zero. So always apply. It also applies to the total weight rather
than just `WeightMinor`.

In `ReadGrams()` I'm converting the weight back from `float64` at the last
minute so that we don't suffer any rounding errors by converting to/from
`int` at each step.