Change fork of gousb for Ubuntu 12.04 support
Ubuntu 12.04, which is what Travis CI runs tests on, has an older version of
libusb which doesn't support `libusb_set_auto_detach_kernel_driver`. This
meant that builds failed with:

    could not determine kind of name for C.libusb_set_auto_detach_kernel_driver

This branch takes yokujin's fix for the warning when compiling on Darwin and
yukkeorg's variation on detach which uses `libusb_detach_kernel_driver`,
plus a commit of my own to not error when detach isn't required.